+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/classification/MaximumEntropyTest.java
@@ -20,7 +20,8 @@ import com.datumbox.framework.common.dataobjects.Dataframe;
 import com.datumbox.framework.common.dataobjects.Record;
 import com.datumbox.framework.core.machinelearning.MLBuilder;
 import com.datumbox.framework.core.machinelearning.modelselection.metrics.ClassificationMetrics;
-import com.datumbox.framework.core.machinelearning.modelselection.validators.KFoldValidator;
+import com.datumbox.framework.core.machinelearning.modelselection.Validator;
+import com.datumbox.framework.core.machinelearning.modelselection.splitters.KFoldSplitter;
 import com.datumbox.framework.tests.Constants;
 import com.datumbox.framework.tests.Datasets;
 import com.datumbox.framework.tests.abstracts.AbstractTest;
@@ -106,7 +107,8 @@ public class MaximumEntropyTest extends AbstractTest {
         MaximumEntropy.TrainingParameters param = new MaximumEntropy.TrainingParameters();
         param.setTotalIterations(10);
 
-        ClassificationMetrics vm = new KFoldValidator<>(ClassificationMetrics.class, conf, k).validate(trainingData, param);
+        ClassificationMetrics vm = new Validator<>(ClassificationMetrics.class, conf)
+                .validate(new KFoldSplitter(k).split(trainingData), param);
         
         double expResult = 0.6051098901098901;
         double result = vm.getMacroF1();