+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/clustering/GaussianDPMMTest.java
@@ -19,7 +19,8 @@ import com.datumbox.framework.common.Configuration;
 import com.datumbox.framework.common.dataobjects.Dataframe;
 import com.datumbox.framework.core.machinelearning.MLBuilder;
 import com.datumbox.framework.core.machinelearning.modelselection.metrics.ClusteringMetrics;
-import com.datumbox.framework.core.machinelearning.modelselection.validators.KFoldValidator;
+import com.datumbox.framework.core.machinelearning.modelselection.Validator;
+import com.datumbox.framework.core.machinelearning.modelselection.splitters.KFoldSplitter;
 import com.datumbox.framework.tests.Constants;
 import com.datumbox.framework.tests.Datasets;
 import com.datumbox.framework.tests.abstracts.AbstractTest;
@@ -108,7 +109,8 @@ public class GaussianDPMMTest extends AbstractTest {
         param.setMu0(new double[]{0.0, 0.0});
         param.setPsi0(new double[][]{{1.0,0.0},{0.0,1.0}});
 
-        ClusteringMetrics vm = new KFoldValidator<>(ClusteringMetrics.class, conf, k).validate(trainingData, param);
+        ClusteringMetrics vm = new Validator<>(ClusteringMetrics.class, conf)
+                .validate(new KFoldSplitter(k).split(trainingData), param);
 
         
         double expResult = 1.0;