+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/topicmodeling/LatentDirichletAllocationTest.java
@@ -21,7 +21,8 @@ import com.datumbox.framework.common.dataobjects.Record;
 import com.datumbox.framework.core.machinelearning.MLBuilder;
 import com.datumbox.framework.core.machinelearning.classification.SoftMaxRegression;
 import com.datumbox.framework.core.machinelearning.modelselection.metrics.ClassificationMetrics;
-import com.datumbox.framework.core.machinelearning.modelselection.validators.KFoldValidator;
+import com.datumbox.framework.core.machinelearning.modelselection.Validator;
+import com.datumbox.framework.core.machinelearning.modelselection.splitters.KFoldSplitter;
 import com.datumbox.framework.core.utilities.text.extractors.UniqueWordSequenceExtractor;
 import com.datumbox.framework.tests.Constants;
 import com.datumbox.framework.tests.abstracts.AbstractTest;
@@ -96,7 +97,8 @@ public class LatentDirichletAllocationTest extends AbstractTest {
         tp.setLearningRate(1.0);
         tp.setTotalIterations(50);
 
-        ClassificationMetrics vm = new KFoldValidator<>(ClassificationMetrics.class, conf, 1).validate(reducedTrainingData, tp);
+        ClassificationMetrics vm = new Validator<>(ClassificationMetrics.class, conf)
+                .validate(new KFoldSplitter(1).split(reducedTrainingData), tp);
         
         double expResult = 0.6843125117743629;
         double result = vm.getMacroF1();