+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Version 0.8.0-SNAPSHOT - Build 20161223
 - The InMemoryConnector now keeps a catalog of every item in a WeakReference Map.
 - In MapDBConfiguration it is now feasible to turn on/off the asynchronous writes.
 - Replaced unnecessary delete() calls with close() when no save() is called.
+- Change the Validation mechanism to support Splitters. The AbstractValidator is now a single Validator class. 
+- K-fold cross validation is now performed by combining the KFoldSplitter and the Validator classes. 
 
 Version 0.7.1-SNAPSHOT - Build 20161217
 ---------------------------------------